new homecontroller, old homecontroller is teamcontroller
