+++ app/views/welcome.scala.html
@@ -26,16 +26,16 @@
             <h2>Why do you see this page?</h2>
 
             <p>
-                The <code>conf/routes</code> file defines a route that tells Play to invoke the <code>HomeController.index</code> action
+                The <code>conf/routes</code> file defines a route that tells Play to invoke the <code>TeamController.index</code> action
                 whenever a browser requests the <code>/</code> URI using the GET method:
             </p>
 
             <pre><code># Home page
-GET     /               controllers.HomeController.index</code></pre>
+                GET     /               controllers.TeamController.index</code></pre>
 
 
             <p>
-                Play has invoked the <code>controllers.HomeController.index</code> method:
+                Play has invoked the <code>controllers.TeamController.index</code> method:
             </p>
 
             <pre><code>public Result index() {
